+++ src/app/globals.css
@@ -21,9 +21,11 @@
 
 html,
 body {
+  box-sizing: border-box;
   padding: 0;
   margin: 0;
   height: 100%;
+  overflow-y: hidden;
 }
 
 body {
@@ -32,11 +34,6 @@ body {
   background: var(--background);
   color: var(--foreground);
   font-family: Arial, Helvetica, sans-serif;
-}
-
-main {
-  display: flex;
-  flex-direction: row;
 }
 
 button {
