+++ src/components/Main.tsx
@@ -3,7 +3,7 @@ import Settings from '@/components/Settings';
 
 export default function Main({ children }: IMain) {
   return (
-    <main className="flex flex-row bg-background text-foreground h-full py-6">
+    <main className="flex flex-row bg-background text-foreground flex-1">
       <Settings />
       {children}
     </main>
